+++ CHANGELOG.md
@@ -75,6 +75,8 @@
 50. Fixed Encoding Issue with Mono API.    (Credits to constfold :3)
 51. Fixed Issue with LoadLibrary failing when a DLL file path contains Non-ASCII Characters.    (Credits to constfold :3)
 52. Fixed Issue with Unicode Command Line.    (Credits to constfold :3)
+53. Removed MelonBase.Compatibility.
+54. Added MelonPlatform Attribute.
 
 ---